refactor so that tests can run without fat jar
