+++ src/test/scala/edu/berkeley/cs/rise/opaque/SharedSparkSessions.scala
@@ -39,6 +39,9 @@ trait MultiplePartitionSparkSession {
     .appName("MultiplePartitionSubquerySuite")
     .config("spark.executor.instances", executorInstances)
     .config("spark.sql.shuffle.partitions", numPartitions)
-    .config("spark.jars", "target/scala-2.12/opaque-assembly-0.1.jar")
+    .config(
+      "spark.jars",
+      "target/scala-2.12/opaque_2.12-0.1.jar,target/scala-2.12/opaque_2.12-0.1-tests.jar"
+    )
     .getOrCreate()
 }
